Rename services to resources & openapi to ogcapi
