+++ baremaps-cli/pom.xml
@@ -105,7 +105,7 @@
     </dependency>
     <dependency>
       <groupId>com.baremaps</groupId>
-      <artifactId>baremaps-openapi</artifactId>
+      <artifactId>baremaps-ogcapi</artifactId>
     </dependency>
     <dependency>
       <groupId>com.baremaps</groupId>
