+++ baremaps-cli/src/main/java/com/baremaps/cli/OpenApi.java
@@ -16,13 +16,13 @@ package com.baremaps.cli;
 
 import static io.servicetalk.data.jackson.jersey.ServiceTalkJacksonSerializerFeature.contextResolverFor;
 
-import com.baremaps.openapi.resources.ApiListingResource;
-import com.baremaps.openapi.resources.CollectionsService;
-import com.baremaps.openapi.resources.ConformanceService;
-import com.baremaps.openapi.resources.RootService;
-import com.baremaps.openapi.resources.StylesService;
-import com.baremaps.openapi.resources.SwaggerResource;
-import com.baremaps.openapi.resources.TilesetsService;
+import com.baremaps.ogcapi.resources.ApiResource;
+import com.baremaps.ogcapi.resources.CollectionsResource;
+import com.baremaps.ogcapi.resources.ConformanceResource;
+import com.baremaps.ogcapi.resources.RootResource;
+import com.baremaps.ogcapi.resources.StylesResource;
+import com.baremaps.ogcapi.resources.SwaggerResource;
+import com.baremaps.ogcapi.resources.TilesetsResource;
 import com.baremaps.postgres.jdbc.PostgresUtils;
 import com.baremaps.server.CorsFilter;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -84,14 +84,14 @@ public class OpenApi implements Callable<Integer> {
     ResourceConfig application =
         new ResourceConfig()
             .registerClasses(
-                ApiListingResource.class,
+                ApiResource.class,
                 SwaggerResource.class,
-                RootService.class,
+                RootResource.class,
                 CorsFilter.class,
-                ConformanceService.class,
-                CollectionsService.class,
-                StylesService.class,
-                TilesetsService.class)
+                ConformanceResource.class,
+                CollectionsResource.class,
+                StylesResource.class,
+                TilesetsResource.class)
             .register(contextResolverFor(mapper))
             .register(
                 new AbstractBinder() {
