+++ baremaps-ogcapi/README.md
@@ -1,0 +1,3 @@
+# baremaps-ogcapi
+
+The `baramaps-ogcapi` module is experimental and not yet functional.
