+++ baremaps-ogcapi/src/main/java/com/baremaps/ogcapi/resources/ApiResource.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.baremaps.openapi.resources;
+package com.baremaps.ogcapi.resources;
 
 import com.google.common.io.Resources;
 import io.swagger.util.Json;
@@ -35,11 +35,11 @@ import javax.ws.rs.core.UriInfo;
 
 @Singleton
 @Path("")
-public class ApiListingResource {
+public class ApiResource {
 
   private static final String OPENAPI = "openapi.yaml";
 
-  public ApiListingResource() {}
+  public ApiResource() {}
 
   public String getVersion() {
     try (InputStream input = Resources.getResource("version.txt").openStream()) {
