+++ baremaps-ogcapi/src/main/resources/initialize_database.sql
