+++ baremaps-ogcapi/src/main/resources/openapi.yaml
