+++ baremaps-ogcapi/src/main/resources/schemas/TileJSON.yaml
