+++ baremaps-ogcapi/src/main/resources/schemas/collection.yaml
