+++ baremaps-ogcapi/src/main/resources/schemas/collections.yaml
