+++ baremaps-ogcapi/src/main/resources/schemas/confClasses.yaml
