+++ baremaps-ogcapi/src/main/resources/schemas/exception.yaml
