+++ baremaps-ogcapi/src/main/resources/schemas/extent.yaml
