+++ baremaps-ogcapi/src/main/resources/schemas/landingPage.yaml
