+++ baremaps-ogcapi/src/main/resources/schemas/link.yaml
