+++ baremaps-ogcapi/src/main/resources/schemas/map-metadata.yaml
