+++ baremaps-ogcapi/src/main/resources/schemas/maps-metadata.yaml
