+++ baremaps-ogcapi/src/main/resources/schemas/mb-layer.yaml
