+++ baremaps-ogcapi/src/main/resources/schemas/mb-style.yaml
