+++ baremaps-ogcapi/src/main/resources/schemas/query.yaml
