+++ baremaps-ogcapi/src/main/resources/schemas/style-set-entry.yaml
