+++ baremaps-ogcapi/src/main/resources/schemas/style-set.yaml
