+++ baremaps-ogcapi/src/main/resources/schemas/vector_layer.yaml
