+++ baremaps-ogcapi/src/main/resources/swagger.html
