+++ baremaps-ogcapi/src/main/resources/version.txt
