--- baremaps-openapi/README.md
@@ -1,3 +1,0 @@
-# baremaps-openapi
-
-The `baramaps-openapi` module is experimental and not yet functional.
