+++ pom.xml
@@ -99,7 +99,7 @@
     <module>baremaps-cli</module>
     <module>baremaps-blob</module>
     <module>baremaps-blob-s3</module>
-    <module>baremaps-openapi</module>
+    <module>baremaps-ogcapi</module>
     <module>baremaps-osm</module>
     <module>baremaps-osm-lmdb</module>
     <module>baremaps-osm-rocksdb</module>
@@ -342,7 +342,7 @@
         <version>${project.parent.version}</version>
       </dependency>
       <dependency>
-        <artifactId>baremaps-openapi</artifactId>
+        <artifactId>baremaps-ogcapi</artifactId>
         <groupId>com.baremaps</groupId>
         <version>${project.parent.version}</version>
       </dependency>
